Lots of changes
-Mostly changing the mpreal wrapper for the boost::multiprecision one.
This will allow me to swap out the mathematica calls for arbitrary
precision boost calls for the binomial probabilities. This is needed
because calling mathematica is just too slow, what a dumb idea. The
boost::multiprecision seems to be a much nicer wrapper, but it is NOT
zero-initialized. Which is a problem using it in an accumulator.
-Also added a --usePMC option to the splitting code, trying to see if this
will beat straight PMC