Changes for building on Windows
-It turns out that graphViz *cannot* be built for 64 bit windows, at least
without significant changes. So a HAS_GRAPHVIZ define is introduced.
-Bugfix for FindGraphViz.cmake
-Added explicit casts to avoid certain warnings
-Changed a round() to a floor()
-Only add graphviz as a dependency to networkReliabilityCommon if it was
found
-Changed type int to std::size_t on certain calls related to binaryDataSet
and derived types
-Fixed a negative value being assigned to an std::size_t, changed to
std::ptrdiff_t
-Added checks to return values of gvLayout, these can fail (in case of
improperly built graphViz library)
-getTreeGraph now attempts to generate the treeGraph, and throws an
exception both if the layout call fails, or if the layout call succeeds
but the graph is not generated (which occurs if graphViz is not present)
-More complicated code to serialize a NetworkReliabilityObsTree - It's not
guaranteed that layout() succeeds, in which case we just save the raw data